changes detailed below were for getting a clean build (free of errors) in Rider, swapping out versions on BouncyCastle and Newtonsoft.Json, and making some minor coding adjustments for the sake of either compile-time or runtime bugs.  details follow:
1) build configuration changes to Frameworks.props were made to make room for new logic for the 'net472' build target.  this way, we leverage BouncyCastle 2.8.9 and Newtonsoft.Json-for-Unity version 13.0.102 for that build.  it must be noted that keeping version continuity between wAI's other dlls and those produced from the build here is what drove the changes.  same for net35.props.

2) most of the changes to the .csproj files were to add a build target of 'net472' to these projects.  adding these build targets was the only way to silence certain errors coming out of Rider, and get the correct versions of BouncyCastle and Newtonsoft.Json into the resultant dlls.  that said, some changes were fully automated and part of Rider's being more helpful than i would prefer.

3) alterations to UnityRpcClient were of two kinds:

3a) to organize the code a bit for readability (since i'm coming back into this code to swap out the calls to UnityWebRequest), and

3b) to add a ReferenceLoopHandling flag to the JsonSerializerSettings class.  this is a change for the sake of the runtime operation of the code, as Unity will not run the prototype script without it (while using the new Newtonsoft.Json library).

testing for these changes took longer than it should because i (frankly) didn't know what i was doing.  really need to brush up on MSBuild details, 'cause we're not in Kansas anymore, Toto.

one other thing to note is that when these dlls are placed into Unity, the Newtonsoft.Json library must be introduced into the project directly as a dll in a Plugins folder.  using the Unity Package Manager to bring in the library results in the absence of serialization at runtime.  this doesn't happen within the other wAI code, and after trying to track it down for a day, i'm calling 'uncle'...

...END-OF-LINE...