adding in the RestfulExtraction code that was taken from RestSharp.
this code hasn't been wired up yet, but once it has been, it'll be
the first step toward replacing the calls to UnityWebRequest and
allowing the wAI-Nethereum solution to have a little security in
its requests to the Jubilee Network.

there may be an easier way to accomplish this, but we'll only figure that
out with time and iteration.

checking this in now, so i can migrate to a new workstation.

...END-OF-LINE...